moved update_xcm_fees, update_xcm_weight to xcm_helper #1053
