+++ pallets/crowdloans/src/mock.rs
@@ -365,7 +365,6 @@ impl crate::Config for Test {
     type MinContribution = MinContribution;
     type MaxVrfs = MaxVrfs;
     type MigrateKeysLimit = MigrateKeysLimit;
-    type UpdateOrigin = EnsureRoot<AccountId>;
     type MigrateOrigin = EnsureRoot<AccountId>;
     type CreateVaultOrigin = CreateVaultOrigin;
     type UpdateVaultOrigin = UpdateVaultOrigin;
@@ -384,6 +383,8 @@ parameter_types! {
 }
 
 impl pallet_xcm_helper::Config for Test {
+    type Event = Event;
+    type UpdateOrigin = EnsureRoot<AccountId>;
     type Assets = Assets;
     type XcmSender = XcmRouter;
     type PalletId = XcmHelperPalletId;
@@ -432,7 +433,7 @@ construct_runtime!(
         DmpQueue: cumulus_pallet_dmp_queue::{Pallet, Call, Storage, Event<T>},
         CumulusXcm: cumulus_pallet_xcm::{Pallet, Event<T>, Origin},
         PolkadotXcm: pallet_xcm::{Pallet, Call, Event<T>, Origin},
-        XcmHelper: pallet_xcm_helper::{Pallet, Storage},
+        XcmHelper: pallet_xcm_helper::{Pallet, Storage, Call, Event<T>},
         XTokens: orml_xtokens::{Pallet, Storage, Call, Event<T>},
     }
 );
@@ -458,7 +459,7 @@ pub(crate) fn new_test_ext() -> sp_io::TestExternalities {
             dot(30f64),
         )
         .unwrap();
-        Crowdloans::update_xcm_fees(Origin::root(), dot(10f64)).unwrap();
+        XcmHelper::update_xcm_fees(Origin::root(), dot(10f64)).unwrap();
     });
 
     ext
@@ -523,7 +524,7 @@ pub fn para_ext(para_id: u32) -> sp_io::TestExternalities {
             dot(30f64),
         )
         .unwrap();
-        Crowdloans::update_xcm_fees(Origin::root(), dot(10f64)).unwrap();
+        XcmHelper::update_xcm_fees(Origin::root(), dot(10f64)).unwrap();
     });
 
     ext
