+++ pallets/liquid-staking/src/mock.rs
@@ -334,6 +334,8 @@ parameter_types! {
 }
 
 impl pallet_xcm_helper::Config for Test {
+    type Event = Event;
+    type UpdateOrigin = UpdateOrigin;
     type Assets = Assets;
     type XcmSender = XcmRouter;
     type PalletId = XcmHelperPalletId;
@@ -349,6 +351,7 @@ parameter_types! {
 
 impl crate::Config for Test {
     type Event = Event;
+    type UpdateOrigin = UpdateOrigin;
     type PalletId = StakingPalletId;
     type SelfParaId = SelfParaId;
     type WeightInfo = ();
@@ -358,7 +361,6 @@ impl crate::Config for Test {
     type AccountIdToMultiLocation = AccountIdToMultiLocation;
     type Assets = Assets;
     type RelayOrigin = RelayOrigin;
-    type UpdateOrigin = UpdateOrigin;
     type UnstakeQueueCapacity = UnstakeQueueCapacity;
     type MinStakeAmount = MinStakeAmount;
     type MinUnstakeAmount = MinUnstakeAmount;
@@ -406,7 +408,7 @@ construct_runtime!(
         DmpQueue: cumulus_pallet_dmp_queue::{Pallet, Call, Storage, Event<T>},
         CumulusXcm: cumulus_pallet_xcm::{Pallet, Event<T>, Origin},
         PolkadotXcm: pallet_xcm::{Pallet, Call, Event<T>, Origin},
-        XcmHelper: pallet_xcm_helper::{Pallet, Storage},
+        XcmHelper: pallet_xcm_helper::{Pallet, Storage, Call, Event<T>},
         XTokens: orml_xtokens::{Pallet, Storage, Call, Event<T>},
     }
 );
@@ -437,7 +439,7 @@ pub(crate) fn new_test_ext() -> sp_io::TestExternalities {
         Assets::mint(Origin::signed(ALICE), DOT, Id(BOB), dot(20000f64)).unwrap();
 
         LiquidStaking::update_staking_pool_capacity(Origin::signed(BOB), dot(10000f64)).unwrap();
-        LiquidStaking::update_xcm_fees(Origin::signed(BOB), dot(10f64)).unwrap();
+        XcmHelper::update_xcm_fees(Origin::signed(BOB), dot(10f64)).unwrap();
     });
 
     ext
@@ -558,7 +560,7 @@ pub fn para_ext(para_id: u32) -> sp_io::TestExternalities {
         Assets::mint(Origin::signed(ALICE), DOT, Id(ALICE), 10000 * DOT_DECIMAL).unwrap();
 
         LiquidStaking::update_staking_pool_capacity(Origin::signed(BOB), dot(10000f64)).unwrap();
-        LiquidStaking::update_xcm_fees(Origin::signed(BOB), dot(10f64)).unwrap();
+        XcmHelper::update_xcm_fees(Origin::signed(BOB), dot(10f64)).unwrap();
     });
 
     ext
