+++ pallets/liquid-staking/src/weights.rs
@@ -60,8 +60,6 @@ pub trait WeightInfo {
     fn set_staking_currency() -> Weight;
     fn update_reserve_factor() -> Weight;
     fn update_staking_pool_capacity() -> Weight;
-    fn update_xcm_fees() -> Weight;
-    fn update_xcm_weight() -> Weight;
     fn add_insurances() -> Weight;
     fn payout_slashed() -> Weight;
     fn on_idle() -> Weight;
@@ -128,16 +126,6 @@ impl<T: frame_system::Config> WeightInfo for SubstrateWeight<T> {
     }
     fn update_staking_pool_capacity() -> Weight {
         (31_000_000 as Weight)
-            .saturating_add(T::DbWeight::get().reads(1 as Weight))
-            .saturating_add(T::DbWeight::get().writes(1 as Weight))
-    }
-    fn update_xcm_fees() -> Weight {
-        (31_000_000 as Weight)
-            .saturating_add(T::DbWeight::get().reads(1 as Weight))
-            .saturating_add(T::DbWeight::get().writes(1 as Weight))
-    }
-    fn update_xcm_weight() -> Weight {
-        (32_000_000 as Weight)
             .saturating_add(T::DbWeight::get().reads(1 as Weight))
             .saturating_add(T::DbWeight::get().writes(1 as Weight))
     }
@@ -221,17 +209,7 @@ impl WeightInfo for () {
             .saturating_add(RocksDbWeight::get().reads(1 as Weight))
             .saturating_add(RocksDbWeight::get().writes(1 as Weight))
     }
-    fn update_xcm_fees() -> Weight {
-        (31_000_000 as Weight)
-            .saturating_add(RocksDbWeight::get().reads(1 as Weight))
-            .saturating_add(RocksDbWeight::get().writes(1 as Weight))
-    }
-    fn update_xcm_weight() -> Weight {
-        (32_000_000 as Weight)
-            .saturating_add(RocksDbWeight::get().reads(1 as Weight))
-            .saturating_add(RocksDbWeight::get().writes(1 as Weight))
-    }
-    fn add_insurances() -> Weight {
+     fn add_insurances() -> Weight {
         (122_000_000 as Weight)
             .saturating_add(RocksDbWeight::get().reads(5 as Weight))
             .saturating_add(RocksDbWeight::get().writes(4 as Weight))
